+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Tap-based calculation of tempo through hitting ENTER key
 - Mute metronome
 - Stop/Start metronome
+- Let user pick which beats are up or down by clicking on them (switching color between red [up] and yellow [down])
 - Show "play logo" triangle inside "ON" beat for accessibility support (e.g. color blind users)
 - Show only 8 beats max horizontally, creating lines of beats if beat count is greater than 8